Unattending should not be possible after the event starts.

+++ apps/events/views.py
@@ -150,6 +150,10 @@ def unattendEvent(request, event_id):
         messages.error(request, _(u"Avmeldingsfristen for dette arrangementet har utløpt."))
         return redirect(event)
 
+    if attendance_Event.start_date < timezone.now():
+        messages.error(request, _(u"Dette arrangementet har allerede startet."))
+        return redirect(event)
+
     event.notify_waiting_list(host=request.META['HTTP_HOST'], unattended_user=request.user)
     Attendee.objects.get(event=attendance_event, user=request.user).delete()
